requirements.txt
fastapi
uvicorn
python-multipart
requests
google-generativeai
groq
matplotlib
sqlmodel
python-dotenv
pytest
pytest-asyncio
httpx
pypdf
reportlab